dat 1AM commit (read description)
Basically wrote the code for calculating time, creating points, and
generating an array of x and y coordinates for the graph. math needs to
be checks (all those parenthesis)...
assuming the code works (which it probably doesn't ;)  ), should just
need to feed the array into a plotpoint function or whatever for the
component.
I'll add documentation at some point for all this - no really i swear i
will.