Input Panel added, Runner made cleaner
I was doing this while you worked on your piece. It is a bit complex,
but it allows the user to input fields for the program to graph. It
utilizing multithreading to pass information from the input panel and to
the ProjectileComponent. Once the user presses the graph button, new
physics values are given to ProjectileComponent and it recalculates its
points. Let me know if you have any questions (as I said before, it
might be a bit complex :).