Fixed Exception, started linking input panel with component
It was one thing (an ArrayOutOfBoundsException). You just needed to
start at 0 in the loop instead of 1. It does not graph correctly,
however. I think you need to factor in the x and y scales into your
calculations. Also, if you wouldn't mind, Projectile Component needs
protected setters for the physics fields and ProjectileInputPanel needs
protected getters for its physics fields. I've written an example of
each. ProjectileRunner needs to utilize these getters and setters when
the main thread is notified. Let me know if you want help with anything.