Made more things customizable
The scales and whether or not the graph time is realistic are now up to
the user. Everything seems to be done EXCEPT the last arrowhead. (I'll
leave that up to you :) Once that is done, we can either show Mr. Wilcox
using an executable jar file or just in eclipse.